requirements.txt
python-redmine
icalendar
python-dotenv